Encapsulate graph traversal logic into MappingGraphAlgorithms
Extracts all core graph traversal methods (Top-Down, Bottom-Up, component group traversals) from `MappingGraph` and moves their implementation to a dedicated `MappingGraphAlgorithms` class. This improves modularity by separating the algorithm details from the main graph structure API.

Updates headers and tests to utilize the new `algorithms` member struct for all traversal calls.
