+++ Sofa/framework/Simulation/Core/src/sofa/simulation/MappingGraph.cpp
@@ -244,123 +244,6 @@ sofa::type::vector<core::BaseMapping*> MappingGraph::getBottomUpMappingsFrom(
     return {};
 }
 
-namespace
-{
-bool shouldVisit(const BaseMappingGraphNode* node, VisitorApplication scope)
-{
-    if (node)
-    {
-        switch (scope)
-        {
-            case VisitorApplication::ONLY_MAPPED_NODES:
-                return node->isMapped();
-            case VisitorApplication::ONLY_MAIN_NODES:
-                return !node->isMapped();
-            case VisitorApplication::ALL_NODES:
-            default:
-                return true; // Visit all nodes.
-        }
-    }
-    return false;
-}
-}
-
-void MappingGraph::traverse(MappingGraphVisitor& visitor, VisitorApplication scope) const
-{
-    for (auto& node : m_allNodes)
-    {
-        if (shouldVisit(node.get(), scope))
-        {
-            node->accept(visitor);
-        }
-    }
-}
-
-void MappingGraph::traverseTopDown(MappingGraphVisitor& visitor, VisitorApplication scope) const
-{
-    std::queue<BaseMappingGraphNode*> ready = prepareRootForTraversal();
-    processQueue(ready, [&visitor, scope](const BaseMappingGraphNode* node)
-    {
-        if (shouldVisit(node, scope))
-        {
-            node->accept(visitor);
-        }
-    });
-}
-
-std::queue<BaseMappingGraphNode*> MappingGraph::prepareRootForTraversal() const
-{
-    std::queue<BaseMappingGraphNode*> ready;
-    for (auto& node : m_allNodes)
-    {
-        node->m_pendingCount = static_cast<int>(node->m_parents.size());
-        if (node->m_pendingCount == 0) //node without any parent -> a root
-        {
-            ready.push(node.get());
-        }
-    }
-
-    return ready;
-}
-
-void MappingGraph::traverseBottomUp(MappingGraphVisitor& visitor, VisitorApplication scope) const
-{
-    //the strategy consists in traversing the graph from top to bottom and
-    //register the traversed nodes in a list. The bottom-up traversal corresponds to the
-    //reversed list.
-
-    sofa::type::vector<BaseMappingGraphNode*> nodes;
-    nodes.reserve(m_allNodes.size());
-    {
-        std::queue<BaseMappingGraphNode*> ready = prepareRootForTraversal();
-        processQueue(ready, [&nodes, scope](BaseMappingGraphNode* node)
-        {
-            if (shouldVisit(node, scope))
-            {
-                nodes.push_back(node);
-            }
-        });
-    }
-
-    for (auto it = nodes.crbegin(); it != nodes.crend(); ++it)
-    {
-        (*it)->accept(visitor);
-    }
-}
-
-void MappingGraph::traverseComponentGroups(MappingGraphVisitor& visitor, VisitorApplication scope) const
-{
-    for (auto& [states, node] : m_groupIndex)
-    {
-        if (shouldVisit(node.get(), scope))
-        {
-            for (auto& child : node->m_children)
-            {
-                child->accept(visitor);
-            }
-        }
-    }
-}
-void MappingGraph::traverseComponentGroups(MappingGraphVisitor& visitor,
-                                            TaskScheduler* taskScheduler) const
-{
-    if (taskScheduler)
-    {
-        sofa::simulation::parallelForEach(*taskScheduler, m_groupIndex.begin(), m_groupIndex.end(),
-            [&visitor](const auto& pair)
-            {
-                for (auto& child : pair.second->m_children)
-                {
-                    child->accept(visitor);
-                }
-            });
-    }
-    else
-    {
-        traverseComponentGroups(visitor);
-    }
-}
-
 bool MappingGraph::isBuilt() const { return m_isBuilt; }
 
 template<class TComponent>
